+++ netring/src/protocol/builtin/mod.rs
@@ -9,6 +9,8 @@ mod icmp;
 mod tcp;
 mod udp;
 
+#[cfg(feature = "dhcp")]
+mod dhcp;
 #[cfg(feature = "dns")]
 mod dns;
 #[cfg(feature = "http")]
@@ -17,10 +19,14 @@ mod http;
 mod kerberos;
 #[cfg(feature = "ldap")]
 mod ldap;
+#[cfg(feature = "netbios-ns")]
+mod nbns;
 #[cfg(feature = "rdp")]
 mod rdp;
 #[cfg(feature = "smb")]
 mod smb;
+#[cfg(feature = "ssdp")]
+mod ssdp;
 #[cfg(feature = "tls")]
 mod tls;
 #[cfg(feature = "tls")]
@@ -30,6 +36,8 @@ pub use icmp::Icmp;
 pub use tcp::Tcp;
 pub use udp::Udp;
 
+#[cfg(feature = "dhcp")]
+pub use dhcp::Dhcp;
 #[cfg(feature = "dns")]
 pub use dns::Dns;
 #[cfg(feature = "http")]
@@ -38,10 +46,14 @@ pub use http::Http;
 pub use kerberos::Kerberos;
 #[cfg(feature = "ldap")]
 pub use ldap::Ldap;
+#[cfg(feature = "netbios-ns")]
+pub use nbns::Nbns;
 #[cfg(feature = "rdp")]
 pub use rdp::Rdp;
 #[cfg(feature = "smb")]
 pub use smb::Smb;
+#[cfg(feature = "ssdp")]
+pub use ssdp::Ssdp;
 #[cfg(feature = "tls")]
 pub use tls::Tls;
 #[cfg(feature = "tls")]